Refactor logger module
gets logging targets from config. Handles only console for now.
